Define framework instances in each framework module
