+++ umongo/frameworks/motor_asyncio.py
@@ -6,6 +6,7 @@ from pymongo.errors import DuplicateKeyError
 import marshmallow as ma
 
 from ..builder import BaseBuilder
+from ..instance import LazyLoaderInstance
 from ..document import DocumentImplementation
 from ..data_objects import Reference
 from ..exceptions import NotCreatedError, UpdateError, DeleteError, NoneReferenceError
@@ -396,3 +397,10 @@ class MotorAsyncIOBuilder(BaseBuilder):
             field.reference_cls = MotorAsyncIOReference
         if isinstance(field, EmbeddedField):
             field.io_validate_recursive = _embedded_document_io_validate
+
+
+class MotorAsyncIOInstance(LazyLoaderInstance):
+    """
+    :class:`umongo.instance.LazyLoaderInstance` implementation for motor-asyncio
+    """
+    BUILDER_CLS = MotorAsyncIOBuilder
