+++ umongo/frameworks/txmongo.py
@@ -6,6 +6,7 @@ from pymongo.errors import DuplicateKeyError
 import marshmallow as ma
 
 from ..builder import BaseBuilder
+from ..instance import LazyLoaderInstance
 from ..document import DocumentImplementation
 from ..data_objects import Reference
 from ..exceptions import NotCreatedError, UpdateError, DeleteError, NoneReferenceError
@@ -341,3 +342,10 @@ class TxMongoBuilder(BaseBuilder):
             field.reference_cls = TxMongoReference
         if isinstance(field, EmbeddedField):
             field.io_validate_recursive = _embedded_document_io_validate
+
+
+class TxMongoInstance(LazyLoaderInstance):
+    """
+    :class:`umongo.instance.LazyLoaderInstance` implementation for txmongo
+    """
+    BUILDER_CLS = TxMongoBuilder
